Explicitly specify dtype for observation space

+++ sc2gym/envs/collect_mineral_shards.py
@@ -90,7 +90,7 @@ class CollectMineralShardsGroupsEnv(BaseMovement2dEnv):
 
     def _get_observation_space(self):
         screen_shape = (2, ) + self.observation_spec[0]["feature_screen"][1:]
-        space = spaces.Box(low=0, high=_PLAYER_RELATIVE_SCALE, shape=screen_shape)
+        space = spaces.Box(low=0, high=_PLAYER_RELATIVE_SCALE, shape=screen_shape, dtype=np.int32)
         return space
 
     def _extract_observation(self, obs):
